Installed rxjs. Refactor in server.ts file
