+++ package.json
@@ -22,7 +22,8 @@
     "@tensorflow/tfjs": "^2.6.0",
     "@tensorflow/tfjs-node": "^2.6.0",
     "express": "^4.16.4",
-    "mongoose": "^5.7.5"
+    "mongoose": "^5.7.5",
+    "rxjs": "^6.6.3"
   },
   "devDependencies": {
     "@types/body-parser": "^1.19.0",
